refactor(clawdstrike-policy-event): extract kinds/sensor_state from edr/receipt/mod.rs
