+++ crates/libs/clawdstrike-policy-event/src/edr/receipt/kinds/mod.rs
@@ -1,5 +1,7 @@
 mod evidence_bundle;
 mod graph_slice;
+mod sensor_state;
 
 pub(crate) use evidence_bundle::*;
 pub(crate) use graph_slice::*;
+pub(crate) use sensor_state::*;
